+++ src/model/Template.ts
@@ -3,7 +3,7 @@ import { Parameter } from './Parameter'
 import { Matrix } from './Matrix'
 import { EmberFunction } from './EmberFunction'
 import { EmberNode } from './EmberNode'
-import { EmberTreeNode } from '../types/types'
+import { NumberedTreeNode } from '../types/types'
 
 export { Template, TemplateImpl }
 
@@ -14,16 +14,16 @@ export { Template, TemplateImpl }
 interface Template extends EmberElement {
 	type: ElementType.Template
 	/** Templated properties. */
-	element?: EmberTreeNode<Parameter | EmberNode | Matrix | EmberFunction>
+	element?: NumberedTreeNode<Parameter | EmberNode | Matrix | EmberFunction>
 	/** Details of the template. */
 	description?: string
 }
 
 class TemplateImpl implements Template {
 	public readonly type: ElementType.Template = ElementType.Template
-	constructor (
+	constructor(
 		public number: number,
-		public element?: EmberTreeNode<Parameter | EmberNode | Matrix | EmberFunction>,
+		public element?: NumberedTreeNode<Parameter | EmberNode | Matrix | EmberFunction>,
 		public description?: string
-	) { }
+	) {}
 }
